refactor: remove playground
