+++ tests/e2e/greptimedbcluster_test.go
@@ -139,6 +139,26 @@ var _ = Describe("Basic test of greptimedb cluster", func() {
 	})
 })
 
+var _ = Describe("Basic test of greptimedb playground", func() {
+	It("Run Playground", func() {
+		var err error
+		go func() {
+			err = playground()
+		}()
+		Expect(err).NotTo(HaveOccurred(), "failed to create playground")
+	})
+})
+
+func playground() error {
+	cmd := exec.Command("../../bin/gtctl", "playground")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		return err
+	}
+	return nil
+}
+
 func createCluster() error {
 	cmd := exec.Command("../../bin/gtctl", "cluster", "create", "mydb", "--timeout", "300")
 	cmd.Stdout = os.Stdout
