--- tests/e2e/greptimedbplayground_test.go
@@ -1,43 +1,0 @@
-// Copyright 2024 Greptime Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package e2e
-
-import (
-	"os"
-	"os/exec"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("Basic test of greptimedb playground", func() {
-	It("Run Playground", func() {
-		var err error
-		go func() {
-			err = playground()
-		}()
-		Expect(err).NotTo(HaveOccurred(), "failed to create playground")
-	})
-})
-
-func playground() error {
-	cmd := exec.Command("../../bin/gtctl", "playground")
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		return err
-	}
-	return nil
-}
