--- tests/e2e/testdata/cluster.yaml
@@ -1,21 +1,0 @@
-cluster:
-  name: mycluster # name of the cluster
-  artifact:
-    version: latest
-  frontend:
-    replicas: 1
-    mysqlAddr: 0.0.0.0:4200
-  datanode:
-    replicas: 3
-    rpcAddr: 0.0.0.0:14100
-    mysqlAddr: 0.0.0.0:14200
-    httpAddr: 0.0.0.0:14300
-  meta:
-    replicas: 1
-    storeAddr: 127.0.0.1:2379
-    serverAddr: 0.0.0.0:3002
-    httpAddr: 0.0.0.0:14001
-
-etcd:
-  artifact:
-    version: v3.5.7
